feat:修改更新版本信息实现
