+++ web/craco.config.js
@@ -1,4 +1,7 @@
 const CracoLessPlugin = require('craco-less');
+const updateVersion = require("./update-version")
+// 构建时期,额外操作, 自动更新版本时间
+updateVersion.updateVersionTime()
 
 module.exports = {
   plugins: [
@@ -14,3 +17,5 @@ module.exports = {
     },
   ],
 };
+
+
